pyproject.toml
[project]
name = "gigly-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-hijack>=3.7.2",
    "psycopg>=3.2.9",
]
